base/p2p_point: change 'parse_addr_port' to include peer id
This commit changes the function 'P2p_point.parse_addr_port' to
'P2p_point.parse_addr_port_id'. We now use `Ocamllex` to parse
correctly an address.

This function now returns a `result` and this implied some changes to
the node. In particular, no exception is raised by the parsing
function anymore. When the user give a wrong IP or a wrong port, it
should be advertised with a proper error message.

To ensure the validity of the configuration we add a configuration
check. If the configuration is invalid, the node fails with an error message.