Tests/python: add black formatter
The black formatter is a so-called opinionated formatter (not a lot of
configuration). This commit adds it to the dev toolkit, reorganizing some build
files in order to make pre-existing linters work in conjunction with black (aka
adapt reported warnings), not against it.

In particular, the Makefile handles all linters, black included, similarly.

The pre-commit-hook also handles git adding files formatted with black