proxy_context: log if ProxyDelegation returns an Error
These Error are ignored, because proxy_context's API is not within
the error monad; and because the calling functions allow it
(for example if raw_get returns an Error, None is returned). However,
even if we can handle these Errors gracefully, they nevertheless indicate that
something wrong is going on. Hence logging them.