lib_shell/test: Add a reference implementation with fuzzy tests
We check using crowbar that the reference implementation which is
obvious to see that it implements the semantics has the same behavior
than the optimized version implemented in the
`Synchronisation_heuristic` module. In quick-check mode, the `list`
combinator seems to generate small lists only. The use a custom
generator which makes much bigger lists.

Co-authored-by: Raphaël Proust <code@bnwr.net>