Stdlib_unix/Lwt_exit: prefix error messages with executable name
This is to avoid alarming messages in children process from causing too
much confusion. In Tezos, and under specific conditions, the node could
attempt to start an external validation process that would crash
immediately causing a new process to start and crash and so on. This
yielded a series of error messages that were not obviously marked as
coming from the child processes.