proto: Update `Voting_period` type, storage and module interface
Currently, the calculation of the voting period assumes that all the voting
periods since the first block of the first protocol version have the same length.
In order to change the `blocks_per_voting_period` constant, we extend the
storage with the current voting period (its index and start level position).