Tests/python: activate_protocol delay
By default, protocol were activated one year in the past (one year delay)
in client.activate_protocol.

Activating in the past is useful for tests which bake blocks manually
using the client.bake command (it avoids waiting). But it's problematic for tests which launch
bakers and expect them to bake with proper delay between block baking
(blocks are baked faster than they are communicated which leads to many
forks).

In particular, test_baker_endorser.py fails spuriously for this reason.

This commit changes the default delay to zero, and updates the test
which bake manually so that activation is done in the past
(this is only needed for performance...).