pre-commit hook: don't fix badly formatted files
So far the pre-commit hook was fixing badly formatted files, by
modifying them in place and `git add`ing the change. However this
does not work, as explained in https://gitlab.com/tezos/tezos/-/issues/996,
if the commit command is of the form `git commit -m "foo" <pathspec>`
(which I believe is unusual, but still, we should support it).

Hence this commit makes the pre-commit hook **check** the formatting
but it does not try to fix it anymore. To mitigate the loss of features,
the hook tells you the command to execute to fix it manually.