Proto/Michelson: backward-compatible hashing
This commit fixes a migration bug related to big-maps indexed by combs
of size >= 4. The optimized representation for combs is different from
the one of Delphi so we cannot use it in places were
backward-compatibility is expected such as computing hashes of big map
keys. We solve the issue by adding a new unparsing mode called
"Hashing" that behaves as "Optimized" except in the case of combs for
which it has the same semantics as Delphi's Optimized.