ci: join opam jobs and changes to wrapper script
The test_wrapper script is now capable of executing dune, opam and shell
tests.

    You must specify scripts/test_wrapper.sh <type> <test> <name>
    Ex: scripts/test_wrapper.sh dune src/lib_base/test lib_base
    Ex: scripts/test_wrapper.sh opam lib_base
    Ex: scripts/test_wrapper.sh shell 'echo "this is a test"' testname

opam tests are now regrouped in one 21 jobs ( groups of 8). Pinning is done at the
beginning of each the job, and a new `remove` action is now executed for
each test.

The unit tests script was adapted to use the new wrapper.