Tests: remove linting from the default test target at root
The target `dune build @runtest_lint`, which is the only one Makefile
and scripts in src/tooling use, still works perfectly. Checking the
linting of the whole codebase inconditionnaly as part of the tests
poisons opam packaging.

Moreover, dune 2 deals with linting itself via the `dune build @lint`
target and we should actually move the whole linting infra to it!