Build/Rust: Add a `rust-toolchain' file
Currently, the documentation of Tezos specifies the supported version
of Rust, and the script used by `make build-deps' enforces that the
correct version of Rust is effectively used.  It also invites
developers with an incorrect version of Rust to set-up the correct
one, and even provides the commands to do it.  There is, however, a
more straightforward way to approach this issue: the `rust-toolchain'.
Adding it at the root of the project is enough.  Modifying it will
also allows all Tezos developers to download and use the new supported
toolchain automatically, without any manual steps to perform.