Proto/Michelson: disallow empty entrypoints in string addresses
Fixes: #643

Issue #643 revealed that addresses with empty entrypoints (like
"addr%") are allowed in Readable notation. Upon PACK, they collapse to
addresses with no ("default") entrypoint (like "addr") thus breaking
injectivity of unparsing.

This change fixes the problem by parsing "addr%" as "addr" using the
string "default" as entrypoint. Non-emptyness of the entrypoint part
of addresses should now be an invariant of the internal
representation.