Proto: update endorsers_per_block and related constants
initial_endorsers is updated so that the it's still 3/4 of
endorsers_per_block; delay_per_missing_endorsement is reduced from 8
seconds to 4 seconds.

The rewards per endorsement are reduced 16 times, 2 times because
blocks are baked two times faster, and 8 times because the
endorsers_per_block value has increased 8 times.

Deposits are 32 times the maximum reward.