Tests/Python: Extend the tests playing with the operation size limit
The purpose of these tests is to check that even with a larger limit
on operation size, the Michelson type checker does not suffer from
stack overflows. To do this, we send it Michelson large values of
various shapes:
- a long byte sequence
- a very nested sequence
- a long list
- a large natural number

In each case, we check that under the limit everything is fine and
above the limit the protocol answers with the expected "Oversized
operaion" error message.

Co-authored-by: Raphaël Cauderlier <raphael.cauderlier@nomadic-labs.com>