Proto/Migration: DFS stitching
Stitch contracts that produce more than one operation.

Those contracts might be affected by the semantic change of switching from
BFS to DFS for inter-contract calls. To guarantee they aren't, we enforce
that the uncontrolled addresses (from parameters but also the SENDER) are
implicit accounts by asserting that they are actually the SOURCE.

Moreover, we sometimes swap the order of operations when two of them
are emmitted to enforce that the uncontrolled address is called last so even
if it performs a DFS call it cannot access the contract in a state that was
inaccessible in BFS.