tezt: Add Proxy tests to RPC_test.ml
The regression files added in this commit look like the ones modified
in the previous commit, with two exceptions:

* The proxy mode warns when --protocol wasn't specified. This happens
  here, because tezt's RPC API doesn't allow to specify --protocol.
* All ./tezos-client commands have the addition of --mode proxy.

For these two reasons, we did not share the regression files between the
two modes. The entire regression files for a protocol weight circa 40Ko.
We hereby thought that not sharing was fine.