proxy server: make it implement /chains/<chain_id>/blocks/<block_id>/protocols
Without that, tezos-clients that query a proxy server always print:

[08:49:57.305] [client2] ./tezos-client --endpoint http://localhost:29932 --base-dir /tmp/tezt-747544/1/client2 rpc get /chains/main/blocks/head/context/contracts/KT1LGwAq4e4cHVFqbsEYo7B59YcfAvdGHk2D/script
[08:49:57.528] [client2] Warning:
[08:49:57.528] [client2]   Failed to acquire the protocol version from the node
[08:49:57.528] [client2]   Did not find service: GET http://localhost:29932/chains/main/blocks/head/protocols

which doesn't look very nice. This commit hence adds support for this
RPC in the proxy server. Note that this is an approximation: when a
node changes protocol, the proxy server in front of it must be updated.
This is not surprising though as proxy servers are specific to a
protocol.