LB/PBT: Introduce so-called Liquidity Baking machines
In this commit, we introduce so-called “machines” to test the
liquidity baking feature.  In this context, machines are transition
systems which allows to run “scenarios”, that is sequences of “steps”
consisting in calling one or more entrypoints of the Exchange
contracts.

There are three key machines implemented in this commit:

   1. The `ConcreteMachine' which runs scenarios on-chain
   2. The `SimulationMachine' which runs scenarios completely
      off-chain
   3. The `ValidationMachine' which runs scenarios against both
      machines

The need for a fast `SimulationMachine' is motivated by the slowness
of the `ConcreteMachine'; the `ValidationMachine' allows to solve the
trustworthiness issue raised by testing a simulation rather than the
actual implementation.

Besides, the `ValidationMachine' checks that the storage of the LB
contracts remain consistent throughout the execution of scenarios,
that is (1) the tez balance stored in the ledger, (2) the storage of
the FA1.2 contract, and (3) the storage of the tzBTC replacement
contract.

Co-author-by: Arvid Jakobsson <arvid.jakobsson@nomadic-labs.com>
Co-author-by: Thomas Letan <lthms@nomadic-labs.com>