Proto/test: Provide a way to configure [cost_per_byte]
The [cost_per_byte] parameter is here to set a price for increasing
the size of the storage; when it comes to testing the blockchain, it
may make the task to predict transaction outcomes harder.  To avoid
this additional challenge when the later is not the focus of a test,
we provide an additional parameter to [Context.init] to set it to
an arbitrary value, (i.e., [Tez.zero] in practice).