Proto: Disable warning as error for older protocols
Because OCaml 4.12 made warning 16 more precise, it now catches
warnings in older protocols.  As a consequence, we have to disable
warning as error for older protocols, which cannot be modified without
modifying their hash.  That being said, we do not want to do that for
future protocols, so we start versioning the `dune_protocol.template'
file.