LB/PBT: Check an attacker cannot get richer by interacting with CPMM
For any sequence of calls from a fixed attacker to the CPMM, the
global effect of the sequence of calls on the balances of the attacker
in tez, tzBTC, and liquidity tokens is such that at least one of the
balances decreases.

We have looked for global-gain attacks using our property-based
testing framework. More precisely, the test consists in generating 100
initial contexts, and for each of these contexts, generating and
executing a sequence of 100 valid contracts calls performed by the
same implicit account `c'. Note that the execution is done with a
subsidy equal to 0, because the objective of this test is to discover
if an attacker can become richer by exploiting the CPMM, not by
depositing funds against liquidity tokens. After each contract call,
the balance of `c' is checked to assert that either one of its
balances (xtz, tzBTC or lqt) has decreased, or they all remained
constant.

Co-author-by: Arvid Jakobsson <arvid.jakobsson@nomadic-labs.com>
Co-author-by: Thomas Letan <lthms@nomadic-labs.com>