Test/Liquidity Baking: use legacy mode for checking the scripts
Because the scripts of Liquidity Baking are originated at migration
time they bypass non-legacy typechecking. If they rely on a feature
that becomes forbidden in a future protocol they should continue to
work fine.

In particular, the SUB instruction (which is used in the scripts)
might become legacy in order to change its return type in the mutez
case.