tezt: Move net and rpc addr cmd arguments from config_init to run
This behaviour change is needed to start nodes using the [datadir]
option. For example, if the given [datadir] is generated from a
snaphsot using [tezos-node snapshot] command line, neither of
[identity_generate] nor [config_init] could be called. In this case,
we need to provide a way to bind net or rpc addresses nevertheless.
So adding those arguments in the run command instead seems better.