Build: Fix `make build-dev-deps'
As of June 22, 2021, opam was picking ocaml-lsp-server.1.6.0.  The
thing is, this version of ocaml-lsp-server has a broken opam file.
More precisely, the ppx_yojson_conv_lib dependency is not constrained
enough.  This is fixed by the 1.6.1 release, so we ask opam to install
at least this version.  This fixes the issue encountered with `make
build-dev-deps' where contributors had to run it twice or thrice for
it to not return an error code.

Fix #1493