Baker: dissociate watching new proposals from preendorsement action
Currently, a baker only starts to watch a proposal once it preendorsed it.
As a consequence, the baker may miss endorsements quorums (and baking
opportunities at the next level).

This commit fixes the behavior to allow a baker to watch PQCs/EQCs events on
the current proposal even if it decided to not preendorse it (eg. the baker
is locked to a different payload hash).