Proto/Michelson: abstract type for Michelson strings
There is an invariant on Michelson strings: they are only allowed to
contain printable characters (either the newline character or a
character whose ascii code is between 32 and 126 inclusive).

Breaking this invariant could result in locked smart contracts: if a
smart contract stores a string that does not respect the invariant its
storage cannot be elaborated anymore so the contract cannot be called.

This commit adds a new module Script_string that maintains the
invariant by representing Michelson strings using an abstract type
Script_string.t. This is similar to the way non-negativity of nat is
maintained in Script_int.