Proto/Michelson: remove a gas hack in a failing case of the CONTRACT instruction
When CONTRACT fails on an existing contract because of a type
mismatch, the gas consumption of the type equality check was reverted
and a hack was used to consume some gas in this case.

This commit enriches the type of the type merging functions to allow
consuming gas even when the types are different. And it removes the
hack.