shell_services: move arbitraries from light mode
The purpose of this commit is to have the arbitraries of the light
mode (for generating values of type merkle_tree and of type raw_context)
lower in the dependency tree, so that they can be reused by more tests.
New tests in the next commits will make us of that.