Proto: inline gas_exhausted_error
The function used to take the context into account to differentiate between block
exhaustion and operation exhaustion, there is only one left now, so there is no
need for the context.
Moreover having a tzresult is confusing when one expects an error.
Inlining it is just the clearest thing to do.