Proto: Refactor existing LB QCheck tests to explicit invariants
We refine the design of the LB machines to add a key parameter to the
[build] and [run] functions, an invariant to check after each blocks.
This allows us to rework our tests, and make them more explicit.
Previously, the [ValidationMachine] had two invariant checks hidden in
[bake].  With this commit, these two invariants checks are turned into
two proper tests.

We also use this opportunity to refine how the tests are display by
Alcotest.