Proto benchmarks and tests: make Proto_alpha snapshotable again
The protocol snapshoting script rewrites "-open" flags in dune files but
not direct uses of "open" in ml and mli files.

This commit makes the proto_alpha directory snapshotable again by
removing all mentions of *_alpha* modules (whose names are not stable)
in ml and mli files and replacing them with "-open" flags in dune
files.

This commit does not modify the protocol, only the protocol-dependent
libraries and tests.