Proto/Michelson: explicit the sharing of the continuation for IIf
Before this commit, `IF { code1 } { code2 } ; code3` was represented
the same way as `IF { code1 ; code3 } { code2 ; code3 }` in the
internal representation. In the former case the code was shared in
memory but it was not possible to account for this sharing when
measuring the size of the instruction sequence.

This commit adds a parameter to the IIf constructor corresponding to
the IF Michelson instruction, to keep track of the shared part `code3`
and only count it once when measuring instruction sequences.