Proto/Michelson: Double type size limit
In previous versions of the protocol, the michelson_maximum_type_size limit was set to 1000 but the contract input types (pair <parameter_type> <storage_type>)
were not checked. Both components, <parameter_type> and
<storage_type> where however checked hence it was possible to build
types as big as 2001.

This commit sets the limit to 2001 to avoid breaking any contract with
both a big storage type and a big parameter type.