Env/Cache: Enforce cache size limit at each insertion
During design, we had decided to only enforce cache size in-between
blocks so that the order of operations inside a block does not matter
with respect to the cache behavior.

That choice is insecure so we rollback to a more standard
implementation where the cache size is enforced at each insertion.

Signed-off-by: Yann Regis-Gianas <yann@nomadic-labs.com>