lib_context: introduce a `Kinded_hash` abstraction in the context dump
Currently the context dump interface ensures that all hashes are
associated with kind information:

```ocaml
type hash = [`Node of Store.hash | `Blob of Store.hash]
```

This is the correct type when storing hashes as children of tree nodes,
but otherwise it's an unnecessary wrapper that introduces a performance
penalty (in particular during snapshot export, which uses these hashes
to track which objects have been visited).

This commit introduces a distinction between plain hashes and kinded
hashes in order to reduce the memory usage of `tree_iteri_unique` during
snapshot creation.