Proto/Michelson: remove the optional entrypoint in ticketer address
This modifies the internal representation of tickets to omit the
entrypoint attached to the address of the ticketer. Before this, we
had the invariant that the entrypoint was always "default", so this
should have no visible effect (except maybe on the cache size of
storages containing tickets). The "default" entrypoint is still
present when tickets are sent from a contract to another.