Env: Fix cache counter inconsistency
When a cache is totally emptied by a block validation, the baker does
not reconstruct the same cache as the validator. Indeed, the validator
inherits its cache from the predecessor block and therefore reuses the
cache counter. By contrast, the baker reloads the cache from the
context and is not able to recover the counter when the domain is
empty. Hence, the baker starts from a zero counter while the validator
starts from a nonzero counter.

The inconsistency between the two counters has been detected by an
integration test.

This commit adds the cache counter in the context so that reloading
the cache from the context correctly initializes the cache counter
even when the context domain is empty. (Notice that, in the previous
version, a single entry was enough to recover the counter.)

This commit also refactors the code to better encapsulate the cache
domain in the Environment_cache module.

Signed-off-by: Yann Regis-Gianas <yann@nomadic-labs.com>