Proto/client_commands/stresstest: fix --single-op-per-pkh-per-block semantic
It used to allow each key to appear only once in the combined set of sources
and destinations of transfers, so that to observe n transactions on top of the
initial block, you needed 2n different keys. It now has the expected effect of
letting each key appear only once as a source, so that only n keys are needed
to get n transactions.