Proto/Baker: remember recent endorsements in the baker's operations pool
Each time a new head is received, the operation_pool field of the state is
cleaned/reset by this function. Instead of emptying it completely, we keep
the endorsements of at most 5 rounds and 1 level in the past, to be able to
include as much endorsements as possible in the next block if this baker is
the proposer. This allows to handle the following situations:

- The baker observes an EQC for (L, R), but a proposal arrived for (L, R+1).
After the flush, extra endorsements on top of (L, R) are 'Branch_refused',
and are not re-sent by the node. If the baker proposes at (L+1, 1), he should
be able to include these extra endorsements. Hence the cache for old rounds.

- The baker receives a head at (L+1, 0) on top of (L, 0), but this head
didn't reach consensus. If the baker who proposes at (L+1, 1) observed some
extra endorsements for (L, 0) that are not included in (L+1, 0), he may want
to add them. But these endorsements become 'Outdated' in the mempool once
(L+1, 0) is received. Hence the cache for previous level.

Fixes https://gitlab.com/tezos/tezos/-/issues/2079 (solution 4 in https://gitlab.com/tezos/tezos/-/issues/2079#note_738359050)

Co-authored-by: iguerNL@Functori <iguer@functori.com>
Co-authored-by: Vincent Botbol <vincent.botbol@nomadic-labs.com>