Micheline: rename encoding cases (titles + descriptions)
The aim of the MR is to avoid special characters (brackets, commas,
etc.) in the `title`s of a union's case. The reason for this is that the
titles can be used in formal, machine-oriented descriptions (field
names, function names, etc.) and sepcial characters are not always
supported by these.