Proto,tx_rollup: Add an operation to originate a tx rollup
This commit adds a new manager operation to proto_alpha:

    Tx_rollup_origination

This operation allows an implicit account to originate a new
transactional rollup. Once created, the tx rollup is assigned a unique
(even in case of reorganizations) address, derived from the operation
hash. This address always starts with the tru1 prefix. For instance,

    tru1HdK6HiR31Xo1bSAr4mwwCek8ExgwuUeHm

is a valid tx rollup address

Co-author-by: Marigold <contact@marigold.dev>
Co-author-by: Nomadic Labs <contact@nomadic-labs.com>