Shell/Prevalidator: avoid costly build of pending ops Set from a Map
Each time `handle_unprocessed` is called to classify some
pending operations (50 with the default parameter), it builds a set
of pending operations hashes from the map of pending operation at the end.
This may be really costly in case there are a lot of operations
(eg. thousands) pending in the mempool.

This commit allows to have the two data structures at the same time. As
demonstrated on Idiazabelnet, this allows to process operations much
faster, decreasing the delay of endorsements reception in our experiments,
that consists in injecting 35k operations, from ~18 seconds to ~8 seconds.