env: versioning of protocol signature
the module type a protocol must implement does not evolve as the rest of
the environment. But for each environment a new module type is
created. This commit remove the equals module type.

The protocol signature version has the same version number as the
environment it got introduce. So there is two protocol signature V0 and
V3