Proto: Generalize the [destination] argument of [Transaction]
This opens the possibility to reuse [Transaction] and the entrypoint
API to interact with layer-2 solutions such as transaction rollups.

Note that this patch is better reviewed using `git show
--ignore-space-change'.  Indeed, it does not add any additional
functionality. Instead, it adds another layer of
abstraction. Unfortunately, the default diff algorithm used by git is
not well suited to demonstrate it, because this patch increases the
indentation level of large section of code.

Co-authored-by: Marigold <contact@marigold.dev>
Co-authored-by: Nomadic Labs <contact@nomadic-labs.com>
Co-authored-by: Oxhead Alpha <info@oxheadalpha.com>