Stress test: add a command that outputs transaction costs
The TPS benchmark is going to need to know costs of various types of
transactions that are used during the stress test. Previously these costs
were hard-coded, however it is not a future-proof solution because any
change in the costs is not guaranteed to be correctly reflected in the
hard-coded values. This change adds a new stress test command that runs the
necessary estimations and then prints the costs in a machine-readable
format (JSON). The estimations are also used by the stress test command
itself instead of hard-coded values.