Plugin,Ithaca2: Refine operation simulation with cache clearance
At protocol activation time, the protocol cache is clear. If a
simulation occurs just before activation and generate a gas
consumption assuming a cache hit, this cache consumption will
be wrong.

With that commit, we correctly empty the predicted cache if
the number of blocks considered in the future sets us after
the activation of the next period.

Signed-off-by: Yann Regis-Gianas <yann@nomadic-labs.com>