Tezt: ~protocols is no longer labeled
The problem is that if you define:

    let register ?x f ~protocols = ...

then you can partially apply it like this:

    register (fun x -> ...)

to get a function of type (protocols:... -> ...), but if you use:

    register @@ fun x -> ...

instead, then you get a compiler error.

This is only true when ~protocols is labeled.
Because we want to add optional arguments to the register function,
I am removing the label on ~protocols.