Relax dependencies constraints for opam-monorepo
To be able to assemble a dune workspace with all of tezos dependencies
opam-monorepo needs to make sure that they all build using dune.

Some packages in the dependency cone don't build with dune yet and to
circumvent this, we maintain dune ports of those packages on a separate
opam repository at
[dune-universe/opam-overlays](https://github.com/dune-universe/opam-overlays).
Each package `pkg.version` that we ported to dune has a
`pkg.version+dune` version on opam-overlays.
Some of the `=` constraints in the tezos repository prevent
opam-monorepo's solver from selecting the dune ports and therefore
prevent it from generating a lockfile.

This PR replaces `=` constraints with pairs of lower and upper bounds
that should allow the selection of a dune port while still ensuring
that the right version is picked.

It also adds an opam extension, x-opam-monorepo-opam-provided that is
interpreted by opam-monorepo. We use it to mark tezos-rust-libs as being
exempt from this dune requirement. opam-monorepo won't vendor it and it
has to be installed via opam, along with the compiler and dune.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>