Proto,tx_rollup: Make `append_message' more predictable wrt. gas
Before this patch, the consumption of gas of `append_message' was hard
to predict, for multiple factor.

  1. The data layout of the inbox was made such that, when appending a
     message, the previous content was fetched. Because of the
     carbonation, each batch within the same block was more expensive.
  2. The first message put in an inbox is responsible for the creation
     of the inbox, and to update the `successor' field of the previous
     inbox… if it exists.

Put together, these two factors were preventing the command to submit
a batch with the tezos client basically useless without setting
manually a gas limit.

This patch improves this with two modifications:

  1. Each message hash is stored in its own storage cell; this means
     you don’t need to fetch the previous content to append
     something. As a nice bonus, this reduces the gas consumption of
     the `Tx_rollup_submit_batch' operation.
  2. The `prepare_metadata' now consumes a certain (fixed) amount
     up-front, then temporarily disable gas consumption.

Put together, these two changes makes the tezos client command usable,
without sacrificing safety as long as the cost model of
`prepare_metadata' is indeed an over-approximation.

Co-authored-by: David Turner <david.turner@oxheadalpha.com>