long_tests: Fix [Long_test.init] call that was not used at the right place
This commit fixes a bug that was introduced when the [init] function was added
to the [Long_test] module.

As [Long_test.update_grafana_dashboard] is using the [config] reference, this
reference needs to be properly initialized in order to be able to reach the
Grafana endpoint, as the adress of this endpoint is stored in the configuration.

Yet, as nothing prevents it, the call to [Long_test.init] that initialises
the [config] reference was done *after* calling [update_grafana_dashboard]
in [tezt/long_tests/main.ml], which led to the following bug:
The Grafana dashboard could not be updated as the Grafana configuration was
not initiailized.