TPS: make the benchmark run with mempool pre-check enabled
This change makes it possible to run the TPS benchmark with the mempool
pre-check enabled by using extra bootstrap accounts that are generated
automatically. This has been tested with 100k accounts and it works smoothly
without taking too long or exhibiting any other limitations. It should be
possible to go even higher in terms of the number of accounts used, however
I have not tested beyond 100k.

It also solves a problem that occurs when regular transactions are mixed
with smart contract transactions in the mempool, i.e. when mempool pre-check
is disabled counters are not always guaranteed to be in the right order.
This used to cause rejected transactions preventing us from seeing the real
TPS results.