Baker/Alpha: use right round and payload hash of transition block
Instead of relying on a dummy payload for the migration block, we instead try to
decode it. If it works, then the new baker happens to use the same data
encodings as the old one (which in particular works for the I->J transition).
If it fails, we default on the previous behavior (aka, having a dedicated dummy
payload for the transition block).

Co-Authored-By: Eugen Zalinescu <eugen.zalinescu@nomadic-labs.com>