Proto/Michelson/IR GADT: Restrict some instructions to operate on non-empty stacks
Without this, we have for example that `SOME` is allowed on the empty
stack (of type (empty_cell, empty_cell) stack_ty) to
produce an (empty_cell option, empty_cell) stack_ty.

This is currently prevented by the stack_type argument of the kinfo
but we want to remove it soon.