CI: remove misc_checks and check_precommit_hook from master pipeline
The policy with regards to master and release branches is not to rerun
tests that have already run in the pipeline of merged branches. This
is the case for misc_checks and check_precommit_hook. Therefore we
remove them from the master pipeline by having them extend
`.test_template`.

To make sure that all test jobs have access to the python
environment (as required for e.g. check_precommit_hook), we load it in
`.test_template`. This should be quick to execute so has no adverse
effect even if the test job in question does not use python.