Build: Upgrade some packages to available versions building with dune
Those packages don't build with dune or don't build correctly in a
duniverse, getting in the way of building tezos with opam-monorepo.

They don't have a dune port because their latest released versions have
been fixed.

If there are good reasons to hold those packages back, we can consider
maintaining a dune port but upgrading them is the prefered solution.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>