Binary-schema: support zero-byte structs
This can happen for trivial encodings such as `tup2 unit unit`, less
trivial ones such as `obj1 (req "kind" (constant "Nil"))` (common when
translating variant types), or when the value is moved to the tag of a
union by the `Compact` encodings (e.g., `Compact.bool`).

Either way, such an encoding used to be represented as a table with a
header but no body — which is invalid in RST — but now simply features a
plain-English message indicating that the value is represented with zero
bytes.