Proto,tx_rollup: Refine how the commitments bound work
Prior to this patch, the number of commitments was limited in a
strange way. The limitation was on the number of finalized
commitments, but it was preventing the creation of unfinalized
commitments.

To improve the UX, we make the limitation about the number of
commitments either finalized or unfinalized.

We use this opportunity to refine a bit the constants values. They are
still subject to change, though.