Proto/Michelson: simplify parse_comparable_data
It is replaced by `parse_data` as it should match its definition already.
Doing so we're reducing code duplication.
The gas cost model was already the same so it should have no impact here.

Downside: it is now behind a stack depth limitation but it should have no impact here either.