bin_node: add '--progress-display-mode' option for progress animation
Problem: It's impossible to track down the progress of the snapshot
import if 'tezos-node snapshot import' prints output to non-TTY
file descriptor. Such functionality can be required in software that wraps
'tezos-node' and displays the progress of the snapshot import
process.

Solution: Add an option that allows choosing whether the progress should
be displayed despite the fact that tezos-node is writing to file instead
of TTY.