Proto/Tests/Helpers: add option to force reveal in managers
The default behaviour to attach a reveal operation to any manager if
the latter is not reveal hinders testing revelations themselves with
integration tests.

We make this behaviour depend on an optional `?force_reveal` argument
to `manager_operation`, and we set it to true for backwards
compatibility. The default value should be false, actually, but here
we avoided breaking existing integration tests and leave fixing this
as follow up in tezos/tezos#2979.

We also patch [manager_operation] to enable/disable force reveal.

We propagate this to all the operation constructors for single manager
operations.

Co-authored-by: Germán Delbianco <german@nomadic-labs.com>
Co-authored-by: vbot <vincent.botbol@nomadic-labs.com>