Proto/Tests: add tests for incorrectly placed and duplicated reveals
Cf #2774

We test that reveal ops succeed only when placed first in the batch.

We extend the test suite to duplicate reveals in batches to show they
fail even if the first one is well placed:

First, we test a batch consisting of duplicate (potentially) valid
reveal operations. We assert the second reveal to fail again with an
[Apply.Incorrect_reveal_position] error, and for the first reveal to
be backtracked.

Then, we test batches with duplicate reveals which follow a failing
one and we assert again the second reveal fails skipped. We do this
for the 3 different reasons a well-placed reveal might fail (as tested
above): gas exhaustion, insolvency, and emptying the balance while
revealing.

Co-authored-by: Germán Delbianco <german@nomadic-labs.com>
Co-authored-by: vbot <vincent.botbol@nomadic-labs.com>