TPS: set max prechecked manager operations count
Previously this value was defaulting to 5000 which was sometimes exceeded
when the benchmark was run with lifted protocol limits. This would cause the
whole benchmark to fail.

To test the change manually, run:

./tezos-tps-evalutaion-benchmark-tps -a blocks-total=1 -v > my-log.txt

Then inspect my-log.txt. You will notice that the RPC call to
/chains/main/mempool/filter succeeds and returns a JSON object with the
correct value of max_prechecked_manager_operations.