Proto,tx_rollup: Add lazy slashing for rejected commitment’s successors
Prior to this patch, only the committers of rejected commitments see
their frozen bond burn.

Commitments that are descendants of rejected commitments are simply
overwritten.  As a consequence, committers of these commitments can
never reclaim their bond.  This is probably safe enough, but there is
a trick: frozen bonds are still part of their stake.

This patch changes that.  Now, prior to updating the layer-1 storage
with a commitment, we check whether one exists already or not.  If so,
we burn their bond once and for all.  As a safety net to avoid edge
cases, we forbid an implicit account to overwrite one of its own
commitment.