Proto,Test: Rewrite test for the cost model
This commit completely removes the previous test and replaces
it with a sequence of tests, one for each case of Michelson
types.

The test is incomplete: it will be completed with more tests
about Michelson type representation (ty) and Michelson code
representation (kinstr) in a forthcoming commit.

The test also fails on purpose: the fixes will be gradually
introduced by forthcoming commits.

Signed-off-by: Yann Regis-Gianas <yann@nomadic-labs.com>