Tx_rollup: Add a client command to request a layer-2 transfer
As of now, the transaction rollup client only provides fairly
low-level commands. One needs to “craft” (produce a JSON) a
transaction, then “craft” a batch, then finally “inject” the
transaction. These functions make sense to test the rollup node with
care, but not that much for end-users.

To improve on the situation, we introduce a new `transfer` command,
that can be used by actual human being.