Tezt,tx_rollup: Fix flakiness of [check_injection]
When the injector flushes its batches of layer-1 operations to inject,
it emits one “injecting” event per layer-1 operation.

[check_injection] is a function that checks that an “injecting” event.

The previous implementation of [check_injection] was flaky, because it
was trying to use the “completed” event emitted once the flush has
ended as a way to determine the absence of the “injecting” event. In
practice, it meant that if a second call to [check_injection] was done
before the “completed” event related to the first “injecting”, then
the test would fail.

To fix this, we prefer to use a timeout to determine whether or not a
“injecting” event is missing or not.