Tx_rollup,Node: remove L2 Genesis block
Removes the necessity to have a special "genesis" block in the L2 chain,
and allows as a side-effect to support inbox/messages that are sent in
the same block as the rollup origination block. The genesis block that
was present before brought some completely unnecessary complexity, just
for the sake of being analogous to L1. This means that now, a rollup
will not have any chain if there aren't any messages that were posted to
it yet.