Tx_rollup,Node: remove need to provide origination block hash
(We can still provide an origination level to bound the crawler).
By forcing to have an explicit block hash for the origination block, we
prevent the rollup node from supporting rollup originations in a block
of round > 0 (in the case where the user does not wait for 2 block
confirmations). Having the origination block be explicit does not bring
any more guarantees than having the level.