Proto,tests: Check operation size limit by default in [Incremental]
This test can be opt-out explicitly, and sometimes it’s OKay. But when
designing protocol extensions that are explicitly limited by the size
of Tezos operations (like Transaction Rollup was), it makes sense to
systematically check the limits.

The good news is: tests of TORU still pass with this change, modulo a
slight change to allow to submit large batches of layer-2 operations
in one layer-1 batch.